fix(): Ajustes do modo de simulação para apresentação dos dados na tela
